Increase the precision of SOG and COG from float to double
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20817 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: e08b9a3ff4074dd4e014a766a1549c5461f1e043

+++ server/wrapper/ivef/de.uniol.inf.is.odysseus.wrapper.ivef/src/de/uniol/inf/is/odysseus/wrapper/ivef/parser/IVEF_0_1_5_Parser.java
@@ -361,11 +361,11 @@ public class IVEF_0_1_5_Parser extends DefaultHandler {
                 obj.setUpdateTimeDR(val);
             }
             else if (key == "SOG") {
-                Float val = Float.parseFloat(value);
+            	Double val = Double.parseDouble(value);
                 obj.setSOG(val);
             }
             else if (key == "COG") {
-                Float val = Float.parseFloat(value);
+            	Double val = Double.parseDouble(value);
                 obj.setCOG(val);
             }
             else if (key == "Lost") {
